backend/requirements.txt
django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
pillow==11.2.1
gunicorn==22.0.0 